Move the patch tests to own test file to make migration easier
The old norm based tets, which will eventually be removed, were
interleaved with the new patch tests.

* The PR will be easier to review and understand as an added file,
  rather than a massively edited old file.

* The existing test file names are vague, in my opinion, making it
  hard to guess where tests of a particular thing are going to be.
  This attempts to improve the naming convention.
