+++ src/serac/physics/tests/CMakeLists.txt
@@ -35,6 +35,7 @@ set(solver_tests
     serac_thermal_solid_solver.cpp
     serac_thermal_functional.cpp
     serac_thermal_solid_functional.cpp
+    test_solid_functional_statics.cpp
     )
 
 serac_add_tests( SOURCES ${solver_tests}
